Initial Bugfixing
Fix most of the bugs in the main program, such as the log not functioning properly due to conflicts with Javacord, the default config not generating properly, and the config not being properly read; however, bugs do still remain with the JFrame. Also added config option to change the discord bot's activity.