Fix logs
Fix a bug where the logger only prints one line and overrides the previous log. Changed the window to update the log when the cursor enters the text area. Change the default text of the log box to make it obvious that when the cursor is moved into the box, the log updates.