Create Functional Dynamic Internal Commands
Create and implement the system for dynamically getting builtin or prebuilt commands. For builtin commands, users can use the predefined-function tag defined as "%int%<InternalCommand>". Users can also create or import external predefined commands in the cmds folder of YamlToBot and by setting the predefined-function tag to "%ext%<YourPredefinedCommand>". Also changed the logger to not include the level of the logged message.